refactor(bridge): separate setupbridge logic
